Reorg: Big commit that organizes the code files into more appropriate folders
