+++ src/client/gui/BlackJackMainGUI.java
@@ -1,4 +1,4 @@
-package src;
+package client.gui;
 
 import javax.swing.*;
 import java.awt.*;
