+++ src/client/gui/ChipRenderer.java
@@ -1,4 +1,4 @@
-
+package client.gui;
 ///  Poker Chips Design
 
 import java.awt.*;
