+++ src/game/Card.java
@@ -1,3 +1,5 @@
+package game;
+
 public class Card {
     private Suit suit;
     private Value value;
