+++ src/game/Suit.java
@@ -1,3 +1,5 @@
+package game;
+
 public enum Suit {
     CLUBS,
     HEARTS,
