+++ src/game/Timer.java
@@ -1,4 +1,4 @@
-package src;
+package game;
 import java.time.Duration;
 import java.time.LocalDateTime;
