+++ src/game/Value.java
@@ -1,4 +1,4 @@
-import java.util.Optional;
+package game;
 
 public enum Value {
     TWO(2, false, 0),
