+++ src/server/ClientThreadWithHooks.java
@@ -1,3 +1,5 @@
+package server;
+
 import networking.Message;
 
 import java.io.IOException;
