+++ src/server/ServerMain.java
@@ -1,3 +1,5 @@
+package server;
+
 public class ServerMain {
     public static void main(String[] args) {
         Server server = new Server(3333);
